// src/enum/permission-keys.enum.ts
export enum PermissionKeys {
  ViewProduct = 'ViewProduct',
  CreateProduct = 'CreateProduct',
  UpdateProduct = 'UpdateProduct',
  DeleteProduct = 'DeleteProduct',
  ViewCategory = 'ViewCategory',
  CreateCategory = 'CreateCategory',
  UpdateCategory = 'UpdateCategory',
  DeleteCategory = 'DeleteCategory',
  ViewReview = 'ViewReview',
  CreateReview = 'CreateReview',
  UpdateReview = 'UpdateReview',
  DeleteReview = 'DeleteReview',
  ViewBuyer = 'ViewBuyer',
  ViewAnyBuyer = 'ViewAnyBuyer',
  CreateBuyer = 'CreateBuyer',
  UpdateBuyer = 'UpdateBuyer',
  DeleteBuyer = 'DeleteBuyer',
  ViewSeller = 'ViewSeller',
  ViewAnySeller = 'ViewAnySeller',
  CreateSeller = 'CreateSeller',
  UpdateSeller = 'UpdateSeller',
  DeleteSeller = 'DeleteSeller',
  ViewCart = 'ViewCart',
  CreateCart = 'CreateCart',
  UpdateCart = 'UpdateCart',
  DeleteCart = 'DeleteCart',
  ViewOrderDetails = 'ViewOrderDetails',
  CreateOrderDetails = 'CreateOrderDetails',
  UpdateOrderDetails = 'UpdateOrderDetails',
  DeleteOrderDetails = 'DeleteOrderDetails',
  ViewPayment = 'ViewPayment',
  CreatePayment = 'CreatePayment',
  UpdatePayment = 'UpdatePayment',
  DeletePayment = 'DeletePayment',
  ViewWishlist = 'ViewWishlist',
  CreateWishlist = 'CreateWishlist',
  UpdateWishlist = 'UpdateWishlist',
  DeleteWishlist = 'DeleteWishlist',
  ViewBuyerAddress = 'ViewBuyerAddress',
  CreateBuyerAddress = 'CreateBuyerAddress',
  UpdateBuyerAddress = 'UpdateBuyerAddress',
  DeleteBuyerAddress = 'DeleteBuyerAddress',
}